Export anchor clicks interceptor for external use
This is required if the anchor click event is stopped or prevented by
some external library before it reaches the document but the developer
wants to forward the event to Abyssa first. Example: an anchor in a
Google Maps popup window. In such case the developer will be able to
attach Abyssa handler to the popup element and catch the event.